feat: update challenge detail UI

+++ lib/screens/challenge_detail/model/challenge_detail_response.dart
@@ -10,7 +10,7 @@ class ChallengeDetailResponse {
   bool? public;
   String? startTime;
   String? endTime;
-  int? prizeAmount;
+  double? prizeAmount;
   String? category;
   String? status;
   String? challengeInfo;
